+++ src/term_image/image/common.py
@@ -47,7 +47,6 @@ from ..exceptions import (
 from ..utils import (
     ClassInstanceMethod,
     ClassProperty,
-    ClassPropertyMeta,
     arg_type_error,
     arg_value_error,
     arg_value_error_msg,
@@ -161,7 +160,7 @@ class Size(Enum):
     ORIGINAL = Hidden()
 
 
-class ImageMeta(ClassPropertyMeta, ABCMeta):
+class ImageMeta(ABCMeta):
     """Type of all render style classes.
 
     NOTE:
@@ -171,8 +170,25 @@ class ImageMeta(ClassPropertyMeta, ABCMeta):
           defined within this package), ``str(cls)`` is equivalent to ``repr(cls)``.
     """
 
+    _forced_support: bool = False
+
     def __str__(self):
         return self.style or super().__str__()
+
+    forced_support = ClassProperty(
+        lambda self: self._forced_support,
+        doc="""Forced render style support
+
+        See the base instance of this metaclass for the complete description.
+        """,
+    )
+
+    @forced_support.setter
+    def forced_support(self, status: bool):
+        if not isinstance(status, bool):
+            raise arg_type_error("forced_support", status)
+
+        self._forced_support = status
 
     style = property(
         cached(
@@ -355,16 +371,19 @@ class BaseImage(metaclass=ImageMeta):
     )
 
     forced_support = ClassProperty(
-        lambda cls: cls._forced_support,
-        doc="""Render style forced support status
+        lambda self: type(self)._forced_support,
+        doc="""Forced render style support
 
         :type: bool
 
         GET:
-            Returns the forced support status of the render style of the invoker.
+            Returns the forced support status of the invoking class or class of the
+            invoking instance.
 
         SET:
-            Forced support is enabled or disabled for the render style of the invoker.
+            Forced support is enabled or disabled for the invoking class.
+
+            Can not be set on an instance.
 
         If forced support is:
 
@@ -373,8 +392,7 @@ class BaseImage(metaclass=ImageMeta):
         * **disabled**, the return value of :py:meth:`is_supported` determines if
           the render style is supported or not.
 
-        By **default**, forced support is **disabled** by the base style class
-        (:py:class:`BaseImage`).
+        By **default**, forced support is **disabled** for all render style classes.
 
         NOTE:
             * This property is :term:`descendant`.
@@ -383,13 +401,6 @@ class BaseImage(metaclass=ImageMeta):
               instantiation of some render style classes.
         """,
     )
-
-    @forced_support.setter
-    def forced_support(cls, status):
-        if not isinstance(status, bool):
-            raise arg_type_error("forced_support", status)
-
-        cls._forced_support = status
 
     frame_duration = property(
         lambda self: self._frame_duration if self._is_animated else None,