+++ CHANGELOG.md
@@ -44,6 +44,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Swapped `N` for `A` in the *method* field of the `iterm2` style-speific render format specification ([#92]).
 - **(BREAKING!)** `term_image.exceptions.StyleError` is now raised instead of style-specific exceptions ([#93]).
 - **(BREAKING!)** `term_image.exceptions.RenderError` is now raised for errors that occur during rendering ([#94]).
+- **(BREAKING!)** `BaseImage.forced_support` can no longer be set via instances ([#95]).
+- **(BREAKING!)** `ITerm2Image.native_anim_max_bytes` can no longer be set or deleted via instances ([#95]).
 
 ### Removed
 - Image scaling ([#88]).
@@ -69,6 +71,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#92]: https://github.com/AnonymouX47/term-image/pull/92
 [#93]: https://github.com/AnonymouX47/term-image/pull/93
 [#94]: https://github.com/AnonymouX47/term-image/pull/94
+[#95]: https://github.com/AnonymouX47/term-image/pull/95
 [08f4e4d]: https://github.com/AnonymouX47/term-image/commit/08f4e4d1514313bbd4278dadde46d21d0b11ed1f
 [fa47742]: https://github.com/AnonymouX47/term-image/commit/fa477424c83474256d4972c4b2cdd4a765bc1cda
 [ed3baa3]: https://github.com/AnonymouX47/term-image/commit/ed3baa38d7621720c007f4662f89d7abadd76ec3